really fix it
We do not have kryo set up to use the `Serializable` interface, so
it `ServiceOptions` being `Serializable` doesn't help, kryo can't
can't serialize it. A one off kryo config just for `ServiceOptions`
seems arbitrary so just avoid serializing it by wrapping it in a
serializable lambda that instantiates it as needed.