scio: make scio-test dependency optional
Do not include scio-test and its dependencies (scalatest etc) in user
production artifacts.

To use `scala-test`, users must now add an explicit dependency. That
seems to already be common practice.

Note that the `ScioOperator` code still contains references to classes
in `scio-test`, but only in methods that are used as part of testing,
like `ScioOperator.mock()`.