remove flo.forking and FLO_FORKING flags
The `flo.forking` flag has become a crutch that is being used to work
around and prolong the pain of serialization issues in both flo and user
code.

It is also confusing to users who believe that `flo.forking` is a flag
that they _should_ be using and are annoyed that it is hard to find
and causes their code to behave differently.