Reload BigQuery job after it is done in order to correctly capture errors.
In the case where a BigQuery job fails immediately (e.g. bucket does not exist in GCS) the job is never reloaded, and errors are lost.
In this case the job is treated as DONE, whilst the returned JobInfo may still report that the job is RUNNING.
By reloading the job after it is DONE, errors are captured and the true state of the job is returned from the library.